refactor(mapkit): extract shared JWT signing utility
Pull the .p8 PKCS#8 parsing and ES256 JWT signing out of the
`/api/mapkit-token` handler into a shared `api/_utils/_mapkit-jwt.ts`
module. The handler now just calls `signMapKitJwt("mapkit-js", ttl)`
and inherits the same diagnostic surface for misconfigured private
keys.

The shared helper supports two JWT kinds:
- `mapkit-js` keeps the optional `origin` claim so browsers must load
  MapKit JS from `MAPKIT_ORIGIN`.
- `maps-server-api` is identical minus the `origin` claim — the
  server-to-server Apple Maps Server API rejects tokens that pin to a
  browser origin.

The imported PKCS8 `CryptoKey` is cached so repeated signing calls
don't repeat the import.
